requirements.txt
# Prod dependencies
singer-python==5.9.0
requests==2.23.0
backoff==1.8.0

# Dev dependencies
black==19.10b0
coverage==5.1
flake8==3.8.3
isort==4.3.21
mypy==0.780
pytest==5.4.3
pytest-socket==0.3.5
requests-mock==1.8.0
